refactor: refactored the code for upload image and create event page
