+++ pages/admin/create-event.js
@@ -8,6 +8,7 @@ import moment from "moment";
 import Cookies from "universal-cookie";
 import { getCookieValue } from "../../lib/cookie";
 import notify from "../../lib/notify";
+import SharedService from "../../services/SharedService";
 
 export async function getServerSideProps(context) {
   try {
@@ -72,6 +73,22 @@ export async function getServerSideProps(context) {
   }
 }
 
+export async function getImageUrl(eventData , response) {
+  return Promise.all([SharedService.uploadImage(eventData.guest_image , {
+    stage: "dev",
+    fileName: eventData.guest_image.name,
+    id: response.data.data.id,
+    category: "events",
+    ContentType: "image/png"
+  }), SharedService.uploadImage(eventData.banner_image , {
+    stage: "dev",
+    fileName: eventData.banner_image.name,
+    id: response.data.data.id,
+    category: "events",
+    ContentType: "image/png"
+  })])
+}
+
 const CreateEvent = () => {
   const cookies = new Cookies();
   const userCookie = cookies.get("userNullcast");
@@ -97,6 +114,8 @@ const CreateEvent = () => {
     return isoDate;
   };
 
+
+
   const createEventHandler = async (e) => {
     const eventData = {
       guest_name: eventDetails.organizerName,
@@ -111,7 +130,7 @@ const CreateEvent = () => {
     };
     try {
       const data = await EventService.createNewEvent(userCookie, eventData);
-      notify(data.data.message);
+      notify("Event Created");
     } catch (error) {
       console.log(error);
     }
